Moved each stage scene to separate classes.
